+++ DomenaManager/Helpers/Payments/Payments.cs
@@ -10,11 +10,11 @@ namespace DomenaManager.Helpers
 {
     public static class Payments
     {
-        public static double CalculateSaldo(int year, Apartment apartment)
+        public static decimal CalculateSaldo(int year, Apartment apartment)
         {
             using (var db = new DB.DomenaDBContext())
             {
-                double paym = db.Payments.Where(x => x.ApartmentId.Equals(apartment.ApartmentId) && x.PaymentRegistrationDate.Year <= year && !x.IsDeleted).Select(x => x.PaymentAmount).DefaultIfEmpty(0).Sum();
+                decimal paym = db.Payments.Where(x => x.ApartmentId.Equals(apartment.ApartmentId) && x.PaymentRegistrationDate.Year <= year && !x.IsDeleted).Select(x => x.PaymentAmount).DefaultIfEmpty(0).Sum();
                 var charg = db.Charges.Include(x => x.Components).Where(x => x.ApartmentId.Equals(apartment.ApartmentId) && x.ChargeDate.Year <= year && !x.IsDeleted);
                 
                 foreach (var ch in charg)
@@ -26,12 +26,12 @@ namespace DomenaManager.Helpers
             }
         }
 
-        public static double CalculateBuildingSaldo(int year, Building building)
+        public static decimal CalculateBuildingSaldo(int year, Building building)
         {
             using (var db = new DB.DomenaDBContext())
             {
                 var apartments = db.Apartments.Where(x => x.BuildingId == building.BuildingId && !x.IsDeleted).Select(x => x.ApartmentId).ToList();
-                double paym = db.Payments.Where(x => apartments.Contains(x.ApartmentId) && x.PaymentRegistrationDate.Year <= year && !x.IsDeleted).Select(x => x.PaymentAmount).DefaultIfEmpty(0).Sum();
+                decimal paym = db.Payments.Where(x => apartments.Contains(x.ApartmentId) && x.PaymentRegistrationDate.Year <= year && !x.IsDeleted).Select(x => x.PaymentAmount).DefaultIfEmpty(0).Sum();
                 var invoices = db.Invoices.Where(x => x.BuildingId == building.BuildingId && x.InvoiceDate.Year <= year && !x.IsDeleted);
 
                 foreach (var inv in invoices)
